:label: [#445] Convert SubmissionProvider to typescript
